Orchestrating roleplaying autonomous AI agents
The main goal of the practical work of this course is to put in practice the basic concepts of agent technology in teams of 4-5 students. We used CrewAI, a framework for orchestrating roleplaying autonomous AI agents. The system aims to simulate a city’s emergency response problem. The scenario includes various
types of agents, which will have to cooperate and coordinate to find the most optimal way to respond to a fire in a city. The ultimate goal is to generate an optimal plan to put the fire out and transport injured people to the nearest hospital.